refactor: organization view feat: decorators and exception handlers
